BE-72, refactor: move bung-related enums to domain package
- Move JoinBungResultEnum and CompleteBungResultEnum from global to domain.bung.enums package
- Update import statements in BungController and BungService
- Remove old enum files from global package
- Simplify exception handling in BungController methods
