+++ src/main/java/io/openur/domain/bung/enums/CompleteBungResultEnum.java
@@ -1,9 +1,9 @@
-package io.openur.global.enums;
+package io.openur.domain.bung.enums;
 
 public enum CompleteBungResultEnum {
     SUCCESSFULLY_COMPLETED("successfully completed"),
-    BUNG_HAS_ALREADY_COMPLETED("bung has already completed"),
-    BUNG_HAS_NOT_STARTED("bung has not started yet"),
+    BUNG_HAS_ALREADY_COMPLETED("You cannot complete bung - bung has already completed"),
+    BUNG_HAS_NOT_STARTED("You cannot complete bung - bung has not started yet"),
     ;
 
     private final String value;
